requirements.txt
flask
gunicorn
numpy
scipy
pyswarm
cvxpy
matplotlib
seaborn
plotly
shap
scikit-learn
flask-cors
requests
pyjwt
stripe
redis
SQLAlchemy
psycopg2  # For PostgreSQL
Flask-Login
Flask-WTF
Flask-Admin
Flask-SQLAlchemy
